classification_train/requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
async-timeout==4.0.2
attrs==22.2.0
basicsr==1.4.2
boto3==1.26.116
botocore==1.29.116
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
datasets==2.11.0
dill==0.3.6
et-xmlfile==1.1.0
facexlib==0.3.0
filelock==3.11.0
filterpy==1.4.5
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gfpgan==1.3.8
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
llvmlite==0.39.1
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
polyleven==0.8
protobuf==3.20.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.9.3
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0.post1
tb-nightly==2.13.0a20230418
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1+cu117
torchmetrics==0.11.4
torchvision==0.14.1+cu117
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
xxhash==3.2.0
yapf==0.33.0
yarl==1.8.2
